# contract/openapi/auth-service.yaml
openapi: 3.1.1
info:
  title: Auth Service
  version: 1.0.0
security:
  - bearerAuth: []
paths:
...
  /auth/user-detail:
    get:
      operationId: getUserDetail
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
  /authorities:
    get:
      operationId: getAuthorities
      parameters:
        - name: page
          in: query
          required: false
          schema:
            default: 0
            type: integer
        - name: size
          in: query
          required: false
          schema:
            default: 20
            type: integer
        - name: sort
          in: query
          required: false
          schema:
            default: attribute ASC
            type: string
        - name: searchField
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityPage'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - authority-controller
    post:
      operationId: addAuthority
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorityData'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityDetail'
          description: Created
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - authority-controller
  /authorities/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteAuthority
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - authority-controller
    get:
      operationId: getAuthority
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - authority-controller
    put:
      operationId: setAuthority
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorityData'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - authority-controller
...
  /users:
    get:
      operationId: getUsers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            default: 0
            type: integer
        - name: size
          in: query
          required: false
          schema:
            default: 20
            type: integer
        - name: sort
          in: query
          required: false
          schema:
            default: email ASC
            type: string
        - required: false
          in: query
          name: searchField
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: attributes
          in: query
          required: false
          schema:
            items:
              type: string
            type: array
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
    post:
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: Created
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: deleteUser
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
    get:
      operationId: getUser
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
  /users/{id}/attributes:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: setAttributes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAttributesData'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
  /users/{id}/authorities:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: setAuthorities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuthoritiesData'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
  /users/{id}/confirm:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: setConfirmed
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BooleanValue'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
  /users/{id}/email:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: setEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEmailData'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
  /users/{id}/enable:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: setEnabled
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BooleanValue'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - user-controller
...
components:
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
  schemas:
    Page:
      type: object
      properties:
        totalElements:
          format: int64
          type: integer
        totalPages:
          format: int32
          type: integer
        page:
          format: int32
          type: integer
        size:
          format: int32
          type: integer
        sort:
          type: string
...
    ErrorCode:
      type: string
      enum:
        - UNKNOWN
        - NOT_FOUND
        - INVALID_BODY
        - INVALID_FIELD
        - INVALID_CAPTCHA
        - INVALID_CREDENTIALS
        - USER_NOT_ENABLED
        - USER_NOT_CONFIRMED
        - EMAIL_ALREADY_EXISTS
        - CANNOT_MANAGE_OWN_ACCOUNT
        - REQUIRED_ATTRIBUTE
    ErrorMessage:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        timestamp:
          format: date-time
          type: string
...
    AttributeValueData:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: string
...
    AttributeValueDetail:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
    AuthorityDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authority:
          type: string
    UserDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        confirmed:
          type: boolean
        enabled:
          type: boolean
        attributes:
          items:
            $ref: '#/components/schemas/AttributeValueDetail'
          type: array
        authorities:
          items:
            $ref: '#/components/schemas/AuthorityDetail'
          type: array
    AuthorityPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              items:
                $ref: '#/components/schemas/AuthorityDetail'
              type: array
    AuthorityData:
      type: object
      required:
        - authority
      properties:
        authority:
          type: string
...
    UserPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              items:
                $ref: '#/components/schemas/UserDetail'
              type: array
    UserData:
      type: object
      required:
        - email
        - confirmed
        - enabled
      properties:
        email:
          type: string
          format: email
        confirmed:
          type: boolean
        enabled:
          type: boolean
    UserAttributesData:
      type: object
      required:
        - attributes
      properties:
        attributes:
          items:
            $ref: '#/components/schemas/AttributeValueData'
          type: array
    UserAuthoritiesData:
      type: object
      required:
        - authorities
      properties:
        authorities:
          items:
            type: string
          type: array
    BooleanValue:
      type: object
      required:
        - value
      properties:
        value:
          type: boolean
    UserEmailData:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
...
  responses:
    client-error:
      description: Client error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    server-error:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'